Fix remaining compilation errors
- Added jakarta.validation dependency to pom.xml
- Fixed all getUser_id() method calls to getUserId()
- Fixed type conversions between int and Long
- Updated repository imports in AuthenticationService
- Fixed method signatures in SearchService
- Resolved all 27 compilation errors

+++ src/main/java/com/socialmediaweb/socialmediaweb/controller/SearchController.java
@@ -21,7 +21,7 @@ public class SearchController {
             @RequestBody SearchCriteriaDTO criteria,
             @RequestParam int userId) {
         try {
-            SearchResponseDTO response = searchService.advancedSearch(criteria, userId);
+            SearchResponseDTO response = searchService.advancedSearch(criteria, (long) userId);
             return ResponseEntity.ok(response);
         } catch (Exception e) {
             return ResponseEntity.status(500).build();
@@ -34,7 +34,7 @@ public class SearchController {
             @RequestParam String query,
             @RequestParam int userId) {
         try {
-            List<SearchResultDTO> results = searchService.quickSearch(query, userId);
+            List<SearchResultDTO> results = searchService.quickSearch(query, (long) userId);
             return ResponseEntity.ok(results);
         } catch (Exception e) {
             return ResponseEntity.status(500).build();
@@ -45,7 +45,7 @@ public class SearchController {
     @GetMapping("/recommendations")
     public ResponseEntity<List<SearchResultDTO>> getRecommendations(@RequestParam int userId) {
         try {
-            List<SearchResultDTO> recommendations = searchService.getRecommendations(userId);
+            List<SearchResultDTO> recommendations = searchService.getRecommendations((long) userId);
             return ResponseEntity.ok(recommendations);
         } catch (Exception e) {
             return ResponseEntity.status(500).build();
@@ -58,7 +58,7 @@ public class SearchController {
         try {
             // For now, return recommendations
             // In future, can be based on follower count, activity, etc.
-            List<SearchResultDTO> trending = searchService.getRecommendations(userId);
+            List<SearchResultDTO> trending = searchService.getRecommendations((long) userId);
             return ResponseEntity.ok(trending);
         } catch (Exception e) {
             return ResponseEntity.status(500).build();
